Tests - Move NWRC network check to own file.
Organizes the test structure a bit.
